Factor out common S3 bits
